requirements.txt
antlr4-python3-runtime>=4.11.1
pydicom>=2.3.0
xnat>=0.4.3